Report empty queues as zero replication lag
Add a callback to the Pulse queue handling for cases when the queue is
found to be empty.  Use that callback to report an empty queue as a
replication lag of zero seconds.  An empty queue means there have
been no new pushes since our last run, therefore we must be caught up.